Decrease MTU to 1200 and double NACK buffer size
This is based on WebRTC which choose a MTU of 1200.

This should improve streaming over tunnels that would encapsulate
packets and push them past the size routers could handle (often 1500).

https://groups.google.com/g/discuss-webrtc/c/gH5ysR3SoZI/m/zrnVHqtUAwAJ

This reduces the amount of bytes that can fit in the NACK buffer by
about 14%, to account for this and keep behavior the same the NACK
buffer has been increased to the next largest size, doubling it.